feat(sha256): Implement
Right now it takes a `std::deque<bool`, this is wasteful.

We may change to use `std::vector<bool>` or because the bits aren't
mutated, to take an `uint32_t*` (or something) and a length; the latter
one would be a bit annoying because we couldn't do `bits[i]` as the
extra padding couldn't be added to the vector (as there is no vector).

This could be worked around using something like Python's
ChainMap (instead of checking all `bits[i]` to see whether is in the
original bits range or in the added padded—which would be a
`std::vector` or `std::array`).